feat: split cypress commands according to the feature
