+++ cypress/e2e/01-admin-login.cy.js
@@ -1,3 +1,5 @@
+import adminLogin from '../support/commands/admin-login'
+
 describe('Admin login', () => {
   beforeEach(() => {
     cy.wait(1000);
