+++ cypress/e2e/03-developer-user-permissions-setup.cy.js
@@ -1,3 +1,6 @@
+import developerUserPermissionsSetup from '../support/commands/developer-user-permission-setup'
+import adminLogin from '../support/commands/admin-login'
+
 describe('Setup developer user role permissions', () => {
   beforeEach(() => {
     cy.wait(1000);
