+++ cypress/e2e/09-add-aggregate-meter-read.cy.js
@@ -1,3 +1,6 @@
+import developerUserLogin from '../support/commands/developer-user-login'
+import addAggregateMeterRead from '../support/commands/add-aggregate-meter-reads'
+
 describe('Add aggregate Meter Read', () => {
     beforeEach(() => {
       cy.wait(1000);
