+++ cypress/e2e/10-add-reservation.cy.js
@@ -1,3 +1,6 @@
+import addReservation from '../support/commands/add-reservation'
+import buyerUserLogin from '../support/commands/buyer-user-login'
+
 describe('Add reservation', () => {
   beforeEach(() => {
     cy.wait(1000);
