+++ cypress/e2e/11-certificate.cy.js
@@ -1,3 +1,6 @@
+import certificate from '../support/commands/certificate'
+import buyerUserLogin from '../support/commands/buyer-user-login'
+
 describe('Certificate Generation', () => {
   beforeEach(() => {
     cy.wait(1000);
