+++ cypress/e2e/12-certificate-filtering.cy.js
@@ -1,3 +1,6 @@
+import certificateFilter from '../support/commands/certificate-filter'
+import buyerUserLogin from '../support/commands/buyer-user-login'
+
 describe('Filter certificates', () => {
     beforeEach(() => {
       cy.wait(1000);
