+++ cypress/support/commands/buyer-user-login.js
@@ -1,0 +1,16 @@
+import 'cypress-file-upload';
+const UI_BASE_URL = Cypress.env('UI_BASE_URL');
+
+Cypress.Commands.add('buyerUserLogin', function () {
+    cy.fixture('buyer-user-login.js').then((data) => {
+      cy.visit(`${UI_BASE_URL}/login`).wait(1000);
+      data.forEach((step) => {
+        if (step.action === 'type') {
+          return cy.get(step.selector).type(step.value);
+        }
+        if (step.action === 'click') {
+          return cy.get(step.selector).click().wait(1000);
+        }
+      });
+    });
+  });
